3ºCARRERA/Proyecto de Sistemas Informáticos/psi-ex/p4/kahootclone/requirements.txt
dj-database-url==0.5.0
Django==3.2.1
django-cors-headers==3.2.1
django-crispy-forms==1.9.2
djangorestframework==3.13
djangorestframework-simplejwt==4
django-htmx==1.7.0
Faker==0.9.1
gunicorn==20.0.4
six==1.15.0
urllib3==1.26.9
whitenoise==5.2.0
coverage==6.4.3
Pillow==9.4.0
# psycopg2-binary contiene ficheros binarios compilados que
# pueden ser incompatibles con algunos sistemas operativos
# si surgen problemas de compatibilidad se puede cambiar la
# "psycopg2-binary==2.9.5" por "psycopg2-binary" 
# Esto es , quitar la versión
psycopg2-binary==2.9.5

python-dotenv==0.21.0